refactor: 중복이미지 제거 및 스타일 수정
