+++ package.json
@@ -51,5 +51,5 @@
     "prettier": "^3.3.3",
     "sass": "^1.80.4",
     "sass-loader": "^16.0.2"
-    }
   }
+}
