+++ src/components/modals/CommonModal.scss
@@ -50,9 +50,9 @@
 
 .modalClose,
 .voteModalBack {
+  cursor: pointer;
   width: 30px;
   height: 30px;
-  cursor: pointer;
 
   &:hover {
     filter: brightness(130%);
@@ -66,6 +66,8 @@
 .voteModalBack {
   @media (max-width: 374px) {
     position: absolute;
+    width: 15px;
+    height: 28px;
     left: 0;
   }
 }
