+++ src/components/myPage/FavoriteIdolCard.scss
@@ -33,8 +33,9 @@
       cursor: pointer;
 
       img {
-        width: 12px;
-        height: 12px;
+        width: 18px;
+        height: 18px;
+        filter: brightness(0);
       }
     }
   }
@@ -89,8 +90,8 @@
         cursor: pointer;
 
         img {
-          width: 8px;
-          height: 8px;
+          width: 15px;
+          height: 15px;
         }
       }
     }
