+++ src/components/myPage/FavoriteIdolList.jsx
@@ -1,7 +1,7 @@
 // FavoriteIdolList.jsx
 import React from "react";
 import FavoriteIdolCard from "./FavoriteIdolCard";
-import deleteButton from "../../assets/images/Vector (1).png";
+import deleteButton from "../../assets/icons/delete.svg";
 import "./FavoriteIdolList.scss";
 
 const FavoriteIdolList = ({
