+++ src/components/myPage/IdolCard.scss
@@ -39,8 +39,8 @@
       align-items: center;
 
       img {
-        width: 35px;
-        height: 25px;
+        width: 63px;
+        height: 45px;
         color: $color-white-1;
       }
     }
